+++ src/main/java/com/shnupbups/extrapieces/ExtraPieces.java
@@ -1,8 +1,10 @@
 package com.shnupbups.extrapieces;
 
 import com.shnupbups.extrapieces.api.EPInitializer;
+import com.shnupbups.extrapieces.blocks.PieceBlock;
 import com.shnupbups.extrapieces.core.PieceSet;
 import com.shnupbups.extrapieces.core.PieceSets;
+import com.shnupbups.extrapieces.core.PieceType;
 import com.shnupbups.extrapieces.core.PieceTypes;
 import com.shnupbups.extrapieces.debug.DebugItem;
 import com.shnupbups.extrapieces.register.ModBlocks;
@@ -15,6 +17,7 @@ import net.fabricmc.fabric.api.event.lifecycle.v1.ServerLifecycleEvents;
 import net.fabricmc.loader.api.FabricLoader;
 import net.minecraft.block.AbstractBlock;
 import net.minecraft.block.Block;
+import net.minecraft.block.Blocks;
 import net.minecraft.item.ItemGroup;
 import net.minecraft.item.ItemStack;
 import net.minecraft.registry.Registries;
@@ -25,7 +28,6 @@ import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
 import java.io.File;
-import java.nio.file.Path;
 
 public class ExtraPieces implements ModInitializer {
     public static final String mod_id = "extrapieces";
@@ -34,127 +36,26 @@ public class ExtraPieces implements ModInitializer {
     public static final String piece_pack_version = "2.9.0";
     public static final Logger logger = LogManager.getFormatterLogger(mod_name);
 
-    public static ItemGroup columnGroup = ItemGroup.create(null, -1)
-            .icon(() ->
-                    new ItemStack(PieceSets.registry.values().stream().findFirst().get().getPieces().get(PieceTypes.COLUMN).asItem())
-            ).displayName(Text.translatable("itemGroup.extrapieces.column"))
-            .entries(
-                    (displayContext, entries) -> {
-                        for (PieceSet set : PieceSets.registry.values()) {
-                            entries.add(set.getPieces().get(PieceTypes.COLUMN).asItem());
+    public static ItemGroup registerItemGroup(String groupId, PieceType type) {
+        return ItemGroup.create(null, -1)
+                .icon(() ->
+                        new ItemStack(PieceSets.registry.get(Blocks.STONE).getPieces().get(type).asItem())
+                ).displayName(Text.translatable("itemGroup.extrapieces." + groupId))
+                .entries(
+                        (displayContext, entries) -> {
+                            if (ModConfigs.columns) {
+                                Registries.BLOCK.forEach(item -> {
+                                    if (item instanceof PieceBlock && ((PieceBlock) item).getType() == type)
+                                        try {
+                                            entries.add(new ItemStack(item, 1));
+                                        } catch (Exception e) {
+                                        }
+                                });
+                            }
                         }
-                    }
-            ).build();
+                ).build();
+    }
 
-    public static ItemGroup slabGroup = ItemGroup.create(null, -1)
-            .icon(() ->
-                    new ItemStack(PieceSets.registry.values().stream().findFirst().get().getPieces().get(PieceTypes.SLAB).asItem())
-            ).displayName(Text.translatable("itemGroup.extrapieces.slab"))
-            .entries(
-                    (displayContext, entries) -> {
-                        for (PieceSet set : PieceSets.registry.values()) {
-                            entries.add(set.getPieces().get(PieceTypes.SLAB).asItem());
-                        }
-                    }
-            ).build();
-
-    public static ItemGroup stairGroup = ItemGroup.create(null, -1)
-            .icon(() ->
-                    new ItemStack(PieceSets.registry.values().stream().findFirst().get().getPieces().get(PieceTypes.STAIRS).asItem())
-            ).displayName(Text.translatable("itemGroup.extrapieces.stairs"))
-            .entries(
-                    (displayContext, entries) -> {
-                        for (PieceSet set : PieceSets.registry.values()) {
-                            entries.add(set.getPieces().get(PieceTypes.STAIRS).asItem());
-                        }
-                    }
-            ).build();
-
-    public static ItemGroup wallGroup = ItemGroup.create(null, -1)
-            .icon(() ->
-                    new ItemStack(PieceSets.registry.values().stream().findFirst().get().getPieces().get(PieceTypes.WALL).asItem())
-            ).displayName(Text.translatable("itemGroup.extrapieces.wall"))
-            .entries(
-                    (displayContext, entries) -> {
-                        for (PieceSet set : PieceSets.registry.values()) {
-                            entries.add(set.getPieces().get(PieceTypes.WALL).asItem());
-                        }
-                    }
-            ).build();
-
-    public static ItemGroup fenceGroup = ItemGroup.create(null, -1)
-            .icon(() ->
-                    new ItemStack(PieceSets.registry.values().stream().findFirst().get().getPieces().get(PieceTypes.FENCE).asItem())
-            ).displayName(Text.translatable("itemGroup.extrapieces.fence"))
-            .entries(
-                    (displayContext, entries) -> {
-                        for (PieceSet set : PieceSets.registry.values()) {
-                            entries.add(set.getPieces().get(PieceTypes.FENCE).asItem());
-                        }
-                    }
-            ).build();
-
-    public static ItemGroup fenceGateGroup = ItemGroup.create(null, -1)
-            .icon(() ->
-                    new ItemStack(PieceSets.registry.values().stream().findFirst().get().getPieces().get(PieceTypes.FENCE_GATE).asItem())
-            ).displayName(Text.translatable("itemGroup.extrapieces.fence_gate"))
-            .entries(
-                    (displayContext, entries) -> {
-                        for (PieceSet set : PieceSets.registry.values()) {
-                            entries.add(set.getPieces().get(PieceTypes.FENCE_GATE).asItem());
-                        }
-                    }
-            ).build();
-
-    public static ItemGroup layerGroup = ItemGroup.create(null, -1)
-            .icon(() ->
-                    new ItemStack(PieceSets.registry.values().stream().findFirst().get().getPieces().get(PieceTypes.LAYER).asItem())
-            ).displayName(Text.translatable("itemGroup.extrapieces.layer"))
-            .entries(
-                    (displayContext, entries) -> {
-                        for (PieceSet set : PieceSets.registry.values()) {
-                            entries.add(set.getPieces().get(PieceTypes.LAYER).asItem());
-                        }
-                    }
-            ).build();
-
-    public static ItemGroup cornerGroup = ItemGroup.create(null, -1)
-            .icon(() ->
-                    new ItemStack(PieceSets.registry.values().stream().findFirst().get().getPieces().get(PieceTypes.CORNER).asItem())
-            ).displayName(Text.translatable("itemGroup.extrapieces.corner"))
-            .entries(
-                    (displayContext, entries) -> {
-                        for (PieceSet set : PieceSets.registry.values()) {
-                            entries.add(set.getPieces().get(PieceTypes.CORNER).asItem());
-                        }
-                    }
-            ).build();
-
-    public static ItemGroup pillarGroup = ItemGroup.create(null, -1)
-            .icon(() ->
-                    new ItemStack(PieceSets.registry.values().stream().findFirst().get().getPieces().get(PieceTypes.POST).asItem())
-            ).displayName(Text.translatable("itemGroup.extrapieces.post"))
-            .entries(
-                    (displayContext, entries) -> {
-                        for (PieceSet set : PieceSets.registry.values()) {
-                            entries.add(set.getPieces().get(PieceTypes.POST).asItem());
-                        }
-                    }
-            ).build();
-
-    public static ItemGroup sidingGroup = ItemGroup.create(null, -1)
-            .icon(() ->
-                    new ItemStack(PieceSets.registry.values().stream().findFirst().get().getPieces().get(PieceTypes.SIDING).asItem())
-            ).displayName(Text.translatable("itemGroup.extrapieces.siding"))
-            .entries(
-                    (displayContext, entries) -> {
-                        for (PieceSet set : PieceSets.registry.values()) {
-                            entries.add(set.getPieces().get(PieceTypes.SIDING).asItem());
-                        }
-                    }
-            ).build();
-
-    
 
     public static File configDir;
     public static File ppDir;
@@ -222,20 +123,17 @@ public class ExtraPieces implements ModInitializer {
             debugLog("EPInitializer " + api.toString());
             api.onInitialize();
         });
-        Block block = new Block(AbstractBlock.Settings.create());
-        Registry.register(Registries.BLOCK, getID("test_block"), block);
 
-
-        ModConfigs.initPiecePacks();
-        ModBlocks.init(PACK);
-
-        ModModels.init(ExtraPieces.PACK);
-//		ModBlocks.finish(PACK);
-//		datapack = Artifice.registerData(getID("ep_data"), new ArtificeResourcePackImpl(ResourceType.SERVER_DATA,null, blah->ModBlocks.init((RuntimeResourcePack) blah)){});
-        RRPCallback.BEFORE_USER.register(a -> a.add(PACK));
+        RRPCallback.BEFORE_USER.register(a -> {
+            ModConfigs.initPiecePacks();
+            ModBlocks.init(PACK);
+            ModModels.init(PACK);
+            a.add(PACK);
+        });
 
         Registry.register(Registries.ITEM, getID("debug_item"), new DebugItem());
-        PACK.dumpDirect(Path.of("C:/Users/kinar/Desktop/ExtraPieces"));
+        ServerLifecycleEvents.SERVER_STARTING.register(server -> {
+        });
         ServerLifecycleEvents.SERVER_STARTED.register(server -> {
             if (ModBlocks.setBuilders.size() != PieceSets.registry.size()) {
                 for (PieceSet.Builder psb : ModBlocks.setBuilders.values()) {
@@ -244,16 +142,35 @@ public class ExtraPieces implements ModInitializer {
                 }
             }
         });
-        Registry.register(Registries.ITEM_GROUP, getID("column_group"), columnGroup);
-        Registry.register(Registries.ITEM_GROUP, getID("slab_group"), slabGroup);
-        Registry.register(Registries.ITEM_GROUP, getID("stair_group"), stairGroup);
-        Registry.register(Registries.ITEM_GROUP, getID("wall_group"), wallGroup);
-        Registry.register(Registries.ITEM_GROUP, getID("fence_group"), fenceGroup);
-        Registry.register(Registries.ITEM_GROUP, getID("fence_gate_group"), fenceGateGroup);
-        Registry.register(Registries.ITEM_GROUP, getID("layer_group"), layerGroup);
-        Registry.register(Registries.ITEM_GROUP, getID("corner_group"), cornerGroup);
-        Registry.register(Registries.ITEM_GROUP, getID("pillar_group"), pillarGroup);
-        Registry.register(Registries.ITEM_GROUP, getID("siding_group"), sidingGroup);
-
+        if (ModConfigs.columns) {
+            Registry.register(Registries.ITEM_GROUP, getID("column_group"), registerItemGroup("column", PieceTypes.COLUMN));
+        }
+        if (ModConfigs.slabs) {
+            Registry.register(Registries.ITEM_GROUP, getID("slab_group"), registerItemGroup("slab", PieceTypes.SLAB));
+        }
+        if (ModConfigs.stairs) {
+            Registry.register(Registries.ITEM_GROUP, getID("stair_group"), registerItemGroup("stair", PieceTypes.STAIRS));
+        }
+        if (ModConfigs.walls) {
+            Registry.register(Registries.ITEM_GROUP, getID("wall_group"), registerItemGroup("wall", PieceTypes.WALL));
+        }
+        if (ModConfigs.fences) {
+            Registry.register(Registries.ITEM_GROUP, getID("fence_group"), registerItemGroup("fence", PieceTypes.FENCE));
+        }
+        if (ModConfigs.fenceGates) {
+            Registry.register(Registries.ITEM_GROUP, getID("fence_gate_group"), registerItemGroup("fence_gate", PieceTypes.FENCE_GATE));
+        }
+        if (ModConfigs.layers) {
+            Registry.register(Registries.ITEM_GROUP, getID("layer_group"), registerItemGroup("layer", PieceTypes.LAYER));
+        }
+        if (ModConfigs.corners) {
+            Registry.register(Registries.ITEM_GROUP, getID("corner_group"), registerItemGroup("corner", PieceTypes.CORNER));
+        }
+        if (ModConfigs.posts) {
+            Registry.register(Registries.ITEM_GROUP, getID("pillar_group"), registerItemGroup("post", PieceTypes.POST));
+        }
+        if (ModConfigs.sidings) {
+            Registry.register(Registries.ITEM_GROUP, getID("siding_group"), registerItemGroup("siding", PieceTypes.SIDING));
+        }
     }
 }