+++ src/main/java/com/shnupbups/extrapieces/ExtraPiecesClient.java
@@ -3,6 +3,8 @@ package com.shnupbups.extrapieces;
 import com.shnupbups.extrapieces.register.ModModels;
 import com.shnupbups.extrapieces.register.ModRenderLayers;
 import net.fabricmc.api.ClientModInitializer;
+import net.fabricmc.fabric.api.client.event.lifecycle.v1.ClientLifecycleEvents;
+import net.fabricmc.fabric.api.client.networking.v1.ClientLoginConnectionEvents;
 
 public class ExtraPiecesClient implements ClientModInitializer {
 //todo: fix this if its broken
@@ -17,7 +19,9 @@ public class ExtraPiecesClient implements ClientModInitializer {
 		/*} catch(Exception e) {
 			ExtraPieces.log("BIG OOF: "+e.getMessage());
 		}*/
-
-		ModRenderLayers.init();
+		ClientLoginConnectionEvents.INIT.register((client,a) -> {
+//			ModModels.init();
+			ModRenderLayers.init();
+		});
 	}
 }