+++ src/main/java/com/shnupbups/extrapieces/blocks/PieceBlockItem.java
@@ -5,6 +5,7 @@ import net.fabricmc.api.Environment;
 import net.minecraft.item.BlockItem;
 import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
+import net.minecraft.text.MutableText;
 import net.minecraft.text.Text;
 import net.minecraft.util.Language;
 
@@ -20,6 +21,7 @@ public class PieceBlockItem extends BlockItem {
 	public PieceBlock getPieceBlock() {
 		return pb;
 	}
+
 
 	@Environment(EnvType.CLIENT)
 	@Override