+++ changelog.md
@@ -1,2 +1,10 @@
+### Version 8.0.2+1.20.1
+- Fixed disabling of blocks.
+- Fixed Pieces Packs not working.
+- Other minor fixes.
+
+### Version 8.0.1+1.20.1
+- Fixed crash on startup
+
 ### Version 8.0.0+1.20.1
 - Initial Port.